test(gateway): split reset lifecycle retry coverage
